# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Corporate Logistics Management System API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Vehicle {
  id: number;
  /** propio | tercero */
  tipo: string;
  modelo: string;
  /** Max weight capacity in kg */
  capacidadPeso: number;
  /** Max volume capacity in m3 */
  capacidadVolumen: number;
  /** gasolina | diesel | gas */
  tipoCombustible: string;
  /** Fuel efficiency in km/liter */
  rendimientoKmLitro: number;
  /** @nullable */
  placa?: string | null;
  createdAt: string;
}

export interface VehicleInput {
  tipo: string;
  modelo: string;
  capacidadPeso: number;
  capacidadVolumen: number;
  tipoCombustible: string;
  rendimientoKmLitro: number;
  placa?: string;
}

export interface VehicleUpdate {
  tipo?: string;
  modelo?: string;
  capacidadPeso?: number;
  capacidadVolumen?: number;
  tipoCombustible?: string;
  rendimientoKmLitro?: number;
  placa?: string;
}

export interface LoadRequirement {
  pesoTotal: number;
  volumenTotal: number;
}

export interface Personnel {
  id: number;
  nombre: string;
  /** chofer | ayudante */
  rol: string;
  /** Daily per-diem rate */
  tarifaViaticos: number;
  /** @nullable */
  telefono?: string | null;
  createdAt: string;
}

export interface PersonnelInput {
  nombre: string;
  rol: string;
  tarifaViaticos: number;
  telefono?: string;
}

export interface PersonnelUpdate {
  nombre?: string;
  rol?: string;
  tarifaViaticos?: number;
  telefono?: string;
}

export interface Sale {
  id: number;
  cliente: string;
  volumenTotal: number;
  pesoTotal: number;
  destino: string;
  /** pendiente | despachado | entregado | cancelado */
  estado: string;
  /** @nullable */
  notas?: string | null;
  createdAt: string;
}

export interface SaleInput {
  cliente: string;
  volumenTotal: number;
  pesoTotal: number;
  destino: string;
  estado?: string;
  notas?: string;
}

export interface SaleUpdate {
  cliente?: string;
  volumenTotal?: number;
  pesoTotal?: number;
  destino?: string;
  estado?: string;
  notas?: string;
}

export interface Dispatch {
  id: number;
  ventaId: number;
  vehiculoId: number;
  choferId: number;
  /** @nullable */
  ayudanteId?: number | null;
  fechaEstimadaSalida: string;
  fechaEstimadaLlegada: string;
  /** @nullable */
  ruta?: string | null;
  /** pre-despacho | aprobado | en-ruta | entregado | cancelado */
  estado: string;
  /** @nullable */
  distanciaKm?: number | null;
  createdAt: string;
  /** @nullable */
  vehiculoModelo?: string | null;
  /** @nullable */
  choferNombre?: string | null;
  /** @nullable */
  ayudanteNombre?: string | null;
  /** @nullable */
  clienteNombre?: string | null;
  /** @nullable */
  destino?: string | null;
}

export interface RoutePoint {
  id: number;
  despachoId: number;
  ubicacion: string;
  orden: number;
  /** @nullable */
  latitud?: number | null;
  /** @nullable */
  longitud?: number | null;
}

export interface TravelCost {
  id: number;
  despachoId: number;
  costoPeajes: number;
  costoCombustible: number;
  costoViaticos: number;
  total: number;
  /** @nullable */
  costoCombustiblePorLitro?: number | null;
}

export interface DispatchDetail {
  id: number;
  ventaId: number;
  vehiculoId: number;
  choferId: number;
  /** @nullable */
  ayudanteId?: number | null;
  fechaEstimadaSalida: string;
  fechaEstimadaLlegada: string;
  /** @nullable */
  ruta?: string | null;
  estado: string;
  /** @nullable */
  distanciaKm?: number | null;
  createdAt: string;
  /** @nullable */
  vehiculoModelo?: string | null;
  /** @nullable */
  choferNombre?: string | null;
  /** @nullable */
  ayudanteNombre?: string | null;
  /** @nullable */
  clienteNombre?: string | null;
  /** @nullable */
  destino?: string | null;
  routePoints?: RoutePoint[];
  costs?: TravelCost;
}

export interface RoutePointInput {
  ubicacion: string;
  orden: number;
  latitud?: number;
  longitud?: number;
}

export interface DispatchInput {
  ventaId: number;
  vehiculoId: number;
  choferId: number;
  ayudanteId?: number;
  fechaEstimadaSalida: string;
  fechaEstimadaLlegada: string;
  ruta?: string;
  distanciaKm?: number;
  routePoints?: RoutePointInput[];
}

export interface DispatchUpdate {
  vehiculoId?: number;
  choferId?: number;
  ayudanteId?: number;
  fechaEstimadaSalida?: string;
  fechaEstimadaLlegada?: string;
  ruta?: string;
  estado?: string;
  distanciaKm?: number;
}

export interface TravelCostInput {
  costoPeajes?: number;
  costoCombustible?: number;
  costoViaticos?: number;
  costoCombustiblePorLitro?: number;
}

export interface CostEstimate {
  costoCombustible: number;
  costoViaticos: number;
  costoPeajes: number;
  total: number;
  dias: number;
  litrosEstimados: number;
  distanciaKm: number;
  costoPorLitro?: number;
}

export interface TollRoute {
  id: number;
  origen: string;
  destino: string;
  cantidadPeajes: number;
  costoTotal: number;
  /** @nullable */
  descripcion?: string | null;
}

export interface TollRouteInput {
  origen: string;
  destino: string;
  cantidadPeajes: number;
  costoTotal: number;
  descripcion?: string;
}

export interface TollRouteUpdate {
  origen?: string;
  destino?: string;
  cantidadPeajes?: number;
  costoTotal?: number;
  descripcion?: string;
}

export interface DashboardSummary {
  totalVehiculos: number;
  vehiculosDisponibles: number;
  vehiculosEnRuta: number;
  ventasPendientes: number;
  despachosPendientes: number;
  despachosEnRuta: number;
  despachosEntregados: number;
  costoTotalEsteMes: number;
}

export interface ScheduleDay {
  fecha: string;
  despachoId: number;
  estado: string;
  destino: string;
}

export interface VehicleScheduleEntry {
  vehiculoId: number;
  modelo: string;
  tipo: string;
  diasOcupados: ScheduleDay[];
}

export interface ActiveDispatch {
  id: number;
  vehiculoId: number;
  vehiculoModelo: string;
  choferNombre: string;
  /** @nullable */
  ayudanteNombre?: string | null;
  destino: string;
  fechaEstimadaSalida?: string;
  fechaEstimadaLlegada: string;
  estado: string;
  /** @nullable */
  ultimaUbicacion?: string | null;
  /** @nullable */
  latitud?: number | null;
  /** @nullable */
  longitud?: number | null;
}

export type ListSalesParams = {
status?: string;
};

export type ListDispatchesParams = {
status?: string;
};

export type GetVehicleScheduleParams = {
/**
 * ISO date string for start of week (defaults to current week)
 */
weekStart?: string;
};