# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Corporate Logistics Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /personnel:
    get:
      operationId: listPersonnel
      tags: [personnel]
      summary: List all personnel
      responses:
        "200":
          description: List of personnel
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Personnel"
    post:
      operationId: createPersonnel
      tags: [personnel]
      summary: Create a personnel record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PersonnelInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Personnel"
  /personnel/{id}:
    get:
      operationId: getPersonnel
      tags: [personnel]
      summary: Get personnel by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Personnel record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Personnel"
        "404":
          description: Not found
    patch:
      operationId: updatePersonnel
      tags: [personnel]
      summary: Update personnel
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PersonnelUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Personnel"
        "404":
          description: Not found
    delete:
      operationId: deletePersonnel
      tags: [personnel]
      summary: Delete personnel
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found

  # --- SALES ---
...
  /dispatches:
    get:
      operationId: listDispatches
      tags: [dispatches]
      summary: List all dispatches
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of dispatches
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Dispatch"
    post:
      operationId: createDispatch
      tags: [dispatches]
      summary: Create a dispatch
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DispatchInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dispatch"
  /dispatches/{id}:
    get:
      operationId: getDispatch
      tags: [dispatches]
      summary: Get a dispatch by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Dispatch with route points and costs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DispatchDetail"
        "404":
          description: Not found
    patch:
      operationId: updateDispatch
      tags: [dispatches]
      summary: Update a dispatch
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DispatchUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dispatch"
        "404":
          description: Not found
    delete:
      operationId: deleteDispatch
      tags: [dispatches]
      summary: Delete a dispatch
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
  /dispatches/{id}/approve:
    post:
      operationId: approveDispatch
      tags: [dispatches]
      summary: Approve a dispatch (changes status to approved/en-route)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Dispatch approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dispatch"
        "404":
          description: Not found
  /dispatches/{id}/costs:
    get:
      operationId: getDispatchCosts
      tags: [dispatches]
      summary: Get travel costs for a dispatch
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Travel costs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TravelCost"
        "404":
          description: Not found
    patch:
      operationId: updateDispatchCosts
      tags: [dispatches]
      summary: Update travel costs for a dispatch
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TravelCostInput"
      responses:
        "200":
          description: Costs updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TravelCost"
        "404":
          description: Not found
  /dispatches/{id}/route-points:
    get:
      operationId: listRoutePoints
      tags: [dispatches]
      summary: List route points for a dispatch
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Route points
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RoutePoint"
    post:
      operationId: addRoutePoint
      tags: [dispatches]
      summary: Add a route point to a dispatch
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoutePointInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoutePoint"
  /dispatches/{dispatchId}/route-points/{pointId}:
    delete:
      operationId: deleteRoutePoint
      tags: [dispatches]
      summary: Remove a route point
      parameters:
        - name: dispatchId
          in: path
          required: true
          schema:
            type: integer
        - name: pointId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found

  # --- VEHICLE RECOMMENDATION ---
...
  /dispatches/{id}/estimate-costs:
    get:
      operationId: estimateDispatchCosts
      tags: [dispatches]
      summary: Estimate costs for a dispatch before approval
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Estimated costs breakdown
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CostEstimate"
        "404":
          description: Not found

  # --- TOLLS ---
...
components:
  schemas:
...
    Personnel:
      type: object
      required: [id, nombre, rol, tarifaViaticos, createdAt]
      properties:
        id:
          type: integer
        nombre:
          type: string
        rol:
          type: string
          description: "chofer | ayudante"
        tarifaViaticos:
          type: number
          description: Daily per-diem rate
        telefono:
          type: ["string", "null"]
        createdAt:
          type: string
    PersonnelInput:
      type: object
      required: [nombre, rol, tarifaViaticos]
      properties:
        nombre:
          type: string
        rol:
          type: string
        tarifaViaticos:
          type: number
        telefono:
          type: string
    PersonnelUpdate:
      type: object
      properties:
        nombre:
          type: string
        rol:
          type: string
        tarifaViaticos:
          type: number
        telefono:
          type: string

    # --- SALES ---
...
    Dispatch:
      type: object
      required: [id, ventaId, vehiculoId, choferId, fechaEstimadaSalida, fechaEstimadaLlegada, estado, createdAt]
      properties:
        id:
          type: integer
        ventaId:
          type: integer
        vehiculoId:
          type: integer
        choferId:
          type: integer
        ayudanteId:
          type: ["integer", "null"]
        fechaEstimadaSalida:
          type: string
        fechaEstimadaLlegada:
          type: string
        ruta:
          type: ["string", "null"]
        estado:
          type: string
          description: "pre-despacho | aprobado | en-ruta | entregado | cancelado"
        distanciaKm:
          type: ["number", "null"]
        createdAt:
          type: string
        # Joined fields
        vehiculoModelo:
          type: ["string", "null"]
        choferNombre:
          type: ["string", "null"]
        ayudanteNombre:
          type: ["string", "null"]
        clienteNombre:
          type: ["string", "null"]
        destino:
          type: ["string", "null"]
    DispatchDetail:
      type: object
      required: [id, ventaId, vehiculoId, choferId, fechaEstimadaSalida, fechaEstimadaLlegada, estado, createdAt]
      properties:
        id:
          type: integer
        ventaId:
          type: integer
        vehiculoId:
          type: integer
        choferId:
          type: integer
        ayudanteId:
          type: ["integer", "null"]
        fechaEstimadaSalida:
          type: string
        fechaEstimadaLlegada:
          type: string
        ruta:
          type: ["string", "null"]
        estado:
          type: string
        distanciaKm:
          type: ["number", "null"]
        createdAt:
          type: string
        vehiculoModelo:
          type: ["string", "null"]
        choferNombre:
          type: ["string", "null"]
        ayudanteNombre:
          type: ["string", "null"]
        clienteNombre:
          type: ["string", "null"]
        destino:
          type: ["string", "null"]
        routePoints:
          type: array
          items:
            $ref: "#/components/schemas/RoutePoint"
        costs:
          $ref: "#/components/schemas/TravelCost"
    DispatchInput:
      type: object
      required: [ventaId, vehiculoId, choferId, fechaEstimadaSalida, fechaEstimadaLlegada]
      properties:
        ventaId:
          type: integer
        vehiculoId:
          type: integer
        choferId:
          type: integer
        ayudanteId:
          type: integer
        fechaEstimadaSalida:
          type: string
        fechaEstimadaLlegada:
          type: string
        ruta:
          type: string
        distanciaKm:
          type: number
        routePoints:
          type: array
          items:
            $ref: "#/components/schemas/RoutePointInput"
    DispatchUpdate:
      type: object
      properties:
        vehiculoId:
          type: integer
        choferId:
          type: integer
        ayudanteId:
          type: integer
        fechaEstimadaSalida:
          type: string
        fechaEstimadaLlegada:
          type: string
        ruta:
          type: string
        estado:
          type: string
        distanciaKm:
          type: number

    # --- ROUTE POINTS ---
    RoutePoint:
      type: object
      required: [id, despachoId, ubicacion, orden]
      properties:
        id:
          type: integer
        despachoId:
          type: integer
        ubicacion:
          type: string
        orden:
          type: integer
        latitud:
          type: ["number", "null"]
        longitud:
          type: ["number", "null"]
    RoutePointInput:
      type: object
      required: [ubicacion, orden]
      properties:
        ubicacion:
          type: string
        orden:
          type: integer
        latitud:
          type: number
        longitud:
          type: number

    # --- TRAVEL COSTS ---
    TravelCost:
      type: object
      required: [id, despachoId, costoPeajes, costoCombustible, costoViaticos, total]
      properties:
        id:
          type: integer
        despachoId:
          type: integer
        costoPeajes:
          type: number
        costoCombustible:
          type: number
        costoViaticos:
          type: number
        total:
          type: number
        costoCombustiblePorLitro:
          type: ["number", "null"]
    TravelCostInput:
      type: object
      properties:
        costoPeajes:
          type: number
        costoCombustible:
          type: number
        costoViaticos:
          type: number
        costoCombustiblePorLitro:
          type: number
    CostEstimate:
      type: object
      required: [costoCombustible, costoViaticos, costoPeajes, total, dias, litrosEstimados, distanciaKm]
      properties:
        costoCombustible:
          type: number
        costoViaticos:
          type: number
        costoPeajes:
          type: number
        total:
          type: number
        dias:
          type: number
        litrosEstimados:
          type: number
        distanciaKm:
          type: number
        costoPorLitro:
          type: number

    # --- TOLL ROUTES ---
...