# lib/api-zod/src/generated/types/costEstimate.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Corporate Logistics Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface CostEstimate {
  costoCombustible: number;
  costoViaticos: number;
  costoPeajes: number;
  total: number;
  dias: number;
  litrosEstimados: number;
  distanciaKm: number;
  costoPorLitro?: number;
}